backend/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-rest-auth
asgiref
django-cors-headers
PyJWT